Separate connection event handlers from socket.js
